Introduce BaseHiveTableLayoutHandle
