+++ presto-hive-common/src/main/java/com/facebook/presto/hive/HivePartition.java
