+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveTableLayoutHandle.java
@@ -24,7 +24,6 @@ import com.facebook.presto.hive.metastore.SemiTransactionalHiveMetastore;
 import com.facebook.presto.hive.metastore.Table;
 import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ConnectorSplit;
-import com.facebook.presto.spi.ConnectorTableLayoutHandle;
 import com.facebook.presto.spi.SchemaTableName;
 import com.facebook.presto.spi.TableNotFoundException;
 import com.facebook.presto.spi.relation.RowExpression;
@@ -47,20 +46,16 @@ import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static java.util.Objects.requireNonNull;
 
 public class HiveTableLayoutHandle
-        implements ConnectorTableLayoutHandle
+        extends BaseHiveTableLayoutHandle
 {
     private final SchemaTableName schemaTableName;
     private final String tablePath;
     private final List<HiveColumnHandle> partitionColumns;
     private final List<Column> dataColumns;
     private final Map<String, String> tableParameters;
-    private final TupleDomain<Subfield> domainPredicate;
-    private final RowExpression remainingPredicate;
     private final Map<String, HiveColumnHandle> predicateColumns;
-    private final TupleDomain<ColumnHandle> partitionColumnPredicate;
     private final Optional<HiveBucketHandle> bucketHandle;
     private final Optional<HiveBucketFilter> bucketFilter;
-    private final boolean pushdownFilterEnabled;
     private final String layoutString;
     private final Optional<Set<HiveColumnHandle>> requestedColumns;
     private final boolean partialAggregationsPushedDown;
@@ -134,18 +129,16 @@ public class HiveTableLayoutHandle
             boolean footerStatsUnreliable,
             Optional<HiveTableHandle> hiveTableHandle)
     {
+        super(domainPredicate, remainingPredicate, pushdownFilterEnabled, partitionColumnPredicate, partitions);
+
         this.schemaTableName = requireNonNull(schemaTableName, "table is null");
         this.tablePath = requireNonNull(tablePath, "tablePath is null");
         this.partitionColumns = ImmutableList.copyOf(requireNonNull(partitionColumns, "partitionColumns is null"));
         this.dataColumns = ImmutableList.copyOf(requireNonNull(dataColumns, "dataColumns is null"));
         this.tableParameters = ImmutableMap.copyOf(requireNonNull(tableParameters, "tableProperties is null"));
-        this.domainPredicate = requireNonNull(domainPredicate, "domainPredicate is null");
-        this.remainingPredicate = requireNonNull(remainingPredicate, "remainingPredicate is null");
         this.predicateColumns = requireNonNull(predicateColumns, "predicateColumns is null");
-        this.partitionColumnPredicate = requireNonNull(partitionColumnPredicate, "partitionColumnPredicate is null");
         this.bucketHandle = requireNonNull(bucketHandle, "bucketHandle is null");
         this.bucketFilter = requireNonNull(bucketFilter, "bucketFilter is null");
-        this.pushdownFilterEnabled = pushdownFilterEnabled;
         this.layoutString = requireNonNull(layoutString, "layoutString is null");
         this.requestedColumns = requireNonNull(requestedColumns, "requestedColumns is null");
         this.partialAggregationsPushedDown = partialAggregationsPushedDown;
@@ -186,17 +179,6 @@ public class HiveTableLayoutHandle
     }
 
     /**
-     * Partitions are dropped when HiveTableLayoutHandle is serialized.
-     *
-     * @return list of partitions if available, {@code Optional.empty()} if dropped
-     */
-    @JsonIgnore
-    public Optional<List<HivePartition>> getPartitions()
-    {
-        return partitions;
-    }
-
-    /**
      * HiveTableHandle is dropped when HiveTableLayoutHandle is serialized.
      *
      * @return HiveTableHandle if available, {@code Optional.empty()} if dropped
@@ -208,27 +190,9 @@ public class HiveTableLayoutHandle
     }
 
     @JsonProperty
-    public TupleDomain<Subfield> getDomainPredicate()
-    {
-        return domainPredicate;
-    }
-
-    @JsonProperty
-    public RowExpression getRemainingPredicate()
-    {
-        return remainingPredicate;
-    }
-
-    @JsonProperty
     public Map<String, HiveColumnHandle> getPredicateColumns()
     {
         return predicateColumns;
-    }
-
-    @JsonProperty
-    public TupleDomain<ColumnHandle> getPartitionColumnPredicate()
-    {
-        return partitionColumnPredicate;
     }
 
     @JsonProperty
@@ -241,12 +205,6 @@ public class HiveTableLayoutHandle
     public Optional<HiveBucketFilter> getBucketFilter()
     {
         return bucketFilter;
-    }
-
-    @JsonProperty
-    public boolean isPushdownFilterEnabled()
-    {
-        return pushdownFilterEnabled;
     }
 
     @JsonProperty
@@ -288,7 +246,7 @@ public class HiveTableLayoutHandle
     @Override
     public Object getIdentifier(Optional<ConnectorSplit> split, PlanCanonicalizationStrategy canonicalizationStrategy)
     {
-        TupleDomain<Subfield> domainPredicate = this.domainPredicate;
+        TupleDomain<Subfield> domainPredicate = this.getDomainPredicate();
 
         // If split is provided, we would update the identifier based on split runtime information.
         if (split.isPresent() && (split.get() instanceof HiveSplit) && domainPredicate.getColumnDomains().isPresent()) {
@@ -309,7 +267,7 @@ public class HiveTableLayoutHandle
         return ImmutableMap.builder()
                 .put("schemaTableName", schemaTableName)
                 .put("domainPredicate", canonicalizeDomainPredicate(domainPredicate, getPredicateColumns(), canonicalizationStrategy))
-                .put("remainingPredicate", canonicalizeRowExpression(remainingPredicate, false))
+                .put("remainingPredicate", canonicalizeRowExpression(this.getRemainingPredicate(), false))
                 .put("constraint", getConstraint(canonicalizationStrategy))
                 // TODO: Decide what to do with bucketFilter when canonicalizing
                 .put("bucketFilter", bucketFilter)
